+++ src/script_editor/__init__.py
@@ -1,3 +1,3 @@
 """Main modules that deal with the Nuke code editor widget."""
-from .nuke_se import NukeScriptEditor
+from .nuke_script_editor import NukeScriptEditor
 from .nuke_se_controllers import CodeEditor
