+++ src/script_editor/nuke_script_editor.py
